Major improvement to dome9 role structure.
Notable changes:
Addition of idepotency by adding a touched file assessing an already
pair machine.
Addition of variable to enable and disable role.

Addition of dome9.yml tasks file so that the main.yml file and
reference the dome9 file if enabled. This also allows me to tag the
role by default.